# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Acetime Business API Documentation",
        "contact": {
            "name": "Acetime Soluções",
            "url": "https://www.acetime.com.br",
            "email": "develop@acetime.com.br"
        },
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/enterprises": {
            "get": {
                "description": "Find All Enterprises",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enterprise"
                ],
                "summary": "Enterprises",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Current page to paginate",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http_exception.HttpSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http_exception.HttpError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Enterprise",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enterprise"
                ],
                "summary": "Create Enterprise",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/enterprises/{id}": {
            "get": {
                "description": "Find All Enterprises",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enterprise"
                ],
                "summary": "Enterprise by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enterprise ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Enterprise",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enterprise"
                ],
                "summary": "Update Enterprise",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EnterpriseID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Enterprise",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enterprise"
                ],
                "summary": "Delete Enterprise",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EnterpriseID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http_exception.HttpError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "any"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "http_exception.HttpSuccess": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "data": {
                    "type": "any"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                }
            }
        }
    }
}